fix(types): use Int64 with checked arithmetic for integer values
Change NInt from Integer (arbitrary precision) to Int64 (64-bit signed)
to match Nix's behavior. Nix uses checked 64-bit integers that throw
errors on overflow rather than wrapping silently.

Changes:
- NAtom: Change NInt constructor from Integer to Int64
- Add checked arithmetic operations (checkedAdd, checkedSub, checkedMul,
  checkedDiv, checkedNeg) that return Either with error messages
- Parser: Add bounds checking for integer literals at parse time
- Exec: Update binary/unary operators to use checked arithmetic
- Builtins: Update addNix, mulNix, divNix to use checked operations
- Convert: Add Int64 FromValue/ToValue instances, fix Integer conversions
- Builtins: Make fromTOML available at top-level scope (not just builtins.fromTOML)

This matches Nix's behavior:
- `9223372036854775807 + 1` → "integer overflow in adding..."
- `9223372036854775807 * 2` → "integer overflow in multiplying..."
- `-9223372036854775808` → "invalid integer '9223372036854775808'"
- `fromTOML "x = 1"` now works (not just `builtins.fromTOML`)

+++ src/Nix/Expr/Shorthands.hs
@@ -25,7 +25,7 @@ mkBool :: Bool -> NExpr
 mkBool = Fix . mkBoolF
 
 -- | Put integer.
-mkInt :: Integer -> NExpr
+mkInt :: Int64 -> NExpr
 mkInt = Fix . mkIntF
 
 -- | Put floating point number.
@@ -232,7 +232,7 @@ mkBoolF :: Bool -> NExprF a
 mkBoolF = NConstant . NBool
 
 -- | Unfixed @mkInt@.
-mkIntF :: Integer -> NExprF a
+mkIntF :: Int64 -> NExprF a
 mkIntF = NConstant . NInt
 
 -- | Unfixed @mkFloat@.
